Спринт 16 Clean Architecture шаблон MVVM
