+++ app/src/main/AndroidManifest.xml
@@ -18,19 +18,19 @@
         android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".player.ui.PlayerDisplayActivity"
+            android:name=".player.ui.activity.PlayerDisplayActivity"
             android:exported="false" />
         <activity
             android:name=".search.ui.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".MediaLibraryActivity"
+            android:name=".mediaLibrary.MediaLibraryActivity"
             android:exported="false" />
         <activity
             android:name=".SettingsActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
