+++ app/src/main/java/com/example/playlistmaker2/player/presentation/model/PlaybackState.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker2.player.presentation.model
+
+class PlaybackState(
+    val playes: Boolean,
+    val position: String,
+)
